Create Lista_Creciente
You can modify a list while iteration over it something that is often a bad idea, but which can help in certain situations. Start with a three-element list of number [10,20,30].Print each number, and then append 5x that number to the end of the list. And when you get to 20, add 100 to the end of the list stop printing (and exit loop) when 5x the current number would be greater than 1000.